Class Diagram
A fost creata diagrama de clase.

+++ src/main/java/ro/mta/se/lab/model/WeatherInfos.java
@@ -55,7 +55,7 @@ public class WeatherInfos {
     public boolean equals(Object obj) {
         WeatherInfos object = (WeatherInfos) obj;
         if(object.getIcon().equals(this.icon) && object.getDescription().equals(this.description)
-           && object.getHumidity() == this.humidity && object.getTemperature() == this.temperature
+           && object.getHumidity() == this.humidity && (int)object.getTemperature() == (int)this.temperature
            && object.getWind() == this.wind){
             return true;
         }
